Move Test file around
