+++ _other_/AppointmentDetailView-test.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import {AppointmentDetailView} from '../../../src/components/appointment'
+import {AppointmentDetailView} from '../src/components/appointment'
 import randerer from 'react-test-renderer';
-import { UpcomingAppointments } from '../../../src/constant'
+import { UpcomingAppointments } from '../src/constant'
 
 describe('Appointment Detail View correctly renderer', () => {
     test('Appointment Detail View renderer without crashing', () => {
